log in modal added, sing up page with image drop box added, routes added to app
